requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
celery==5.4.0
python-multipart==0.0.9
PyJWT==2.9.0
redis==5.0.8
